+++ Cliente.js
@@ -1,0 +1,5 @@
+export class Cliente{
+    nome;
+    cpf;
+    rg;
+}
